Here’s a commit message you can use:
```
docs(site): update supported formats, platforms, and version to 0.11.0

- Updated landing and docs pages to reflect all supported platforms (Android, Java, Kotlin/JVM, Spring Boot, KMP)
- Expanded supported .env formats table and added Kotlin type column
- Added Supported Platforms section to docs
- Fixed generated code preview (int → long)
- Updated version references from 0.10.0 to 0.11.0

```

+++ site/src/components/CodePreviewSection.tsx
@@ -60,9 +60,9 @@ DEBUG_ENABLED=true`}
                 <span className="text-on-surface-variant">;</span>
                 {"\n"}
                 {"    "}
-                <span className="text-primary">public static final int</span>
+                <span className="text-primary">public static final long</span>
                 <span className="text-on-surface-variant"> MAX_RETRIES = </span>
-                <span className="text-tertiary">3</span>
+                <span className="text-tertiary">3L</span>
                 <span className="text-on-surface-variant">;</span>
                 {"\n"}
                 {"    "}
